sync/requirements.txt
requests>=2.31.0
python-dotenv>=1.0.1
fritzconnection>=1.10.3